SPA mode -> slow clientLoader in home route -> how to get a loading state?
docs say

https://reactrouter.com/start/framework/data-loading#client-data-loading

```
// HydrateFallback is rendered while the client loader is running
export function HydrateFallback() {
  return <div>Loading...</div>;
}
```

and

https://reactrouter.com/how-to/client-data

```
Export clientLoader and a HydrateFallback when you want to use client data
```

but in SPA mode HydrateFallback is forbidden.

also, when there is no loader, clientLoader is automatically added
hydrate=true, so the route component is not rendered until clientLoader
has finished.

so how can we access the loading state of a route with a slow
clientLoader, in SPA mode??